Updated launch, stop, and running to return model IDs
The ‘launch’, ‘stop’, and ‘running’ functions were updated to return the model_id for their respective actions. ‘launch’ returns the ID of the launched model, ‘running’ returns a list of IDs for currently running models, and ‘stop’ returns the ID of the stopped model or a list of IDs for multiple stopped models.
Additionally, a new ‘info’ endpoint was added to ‘running’ endpoint to provide extra details about the running models.