WebSocket Updates and Enhanced Draw Block Functionality
Improved all WebSocket asynchronous blocks to wait for a response before proceeding. In case of an error, the response will be -1. Additionally, the draw block combines the costumes from the main image costume with saved pen trails costume to generate a new custom output containing both the detections and the sent image.