Extended The Class To Support Custom ONNX Model
Refactored the launch method to differentiate between predefined Jetson-Inference models and custom ONNX models. If a user specifies a custom model, the code now attempts to load it from /usr/local/bin/networks. When a custom model is used, images are resized to 300×300 during inference to match the model’s expected resolution. Added an is_custom property and to identified the customs lauched models.